Create folder for components, move components
