+++ src/app/components/book-details/book-details.component.html
@@ -12,7 +12,9 @@
 
     <fluent-price>{{ book.price | currency }}</fluent-price>
   </fleunt-quantity>
-  <fluent-button appearance="outline">Add to cart</fluent-button>
+  <fluent-button appearance="outline" (click)="addToCart(book)"
+    >Add to cart</fluent-button
+  >
   <button [routerLink]="['']">Back</button>
   <!-- button 'Back' is temporary -->
 </fluent-card>
