+++ src/app/components/book-details/book-details.component.scss
