+++ src/app/components/book-details/book-details.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
-import { BookData } from '../data';
-import { BookService } from '../services/book.service';
+import { BookData } from '../../data';
+import { BookService } from '../../services/book.service';
+import { CartService } from '../../services/cart.service';
 import { ActivatedRoute } from '@angular/router';
 
 @Component({
@@ -12,8 +13,14 @@ export class BookDetailsComponent {
   public book: BookData | undefined;
   constructor(
     private bookService: BookService,
+    private cartService: CartService,
     private route: ActivatedRoute
   ) {}
+
+  addToCart(book: BookData) {
+    this.cartService.addToCart(book);
+    // window.alert('You have added product to the cart');
+  }
 
   getBooks(): void {
     let bookId = parseInt(this.route.snapshot.params['id']);
