+++ src/app/components/book-list/book-list.component.scss
