+++ src/app/components/top-navigation/top-navigation.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import data from '../data.json';
-import { AllData } from '../data';
+import data from '../../data.json';
+import { AllData } from '../../data';
 @Component({
   selector: 'app-top-navigation',
   templateUrl: './top-navigation.component.html',
